Resolve feedbacks
