+++ src/main/java/org/gradle/profiler/BuildToolCommandLineInvoker.java
@@ -2,8 +2,11 @@ package org.gradle.profiler;
 
 import org.gradle.profiler.result.BuildInvocationResult;
 
+import java.io.File;
+import java.time.Duration;
 import java.util.List;
 import java.util.function.Consumer;
+import java.util.function.Supplier;
 
 import static org.gradle.profiler.Phase.MEASURE;
 import static org.gradle.profiler.Phase.WARM_UP;
@@ -26,5 +29,29 @@ public abstract class BuildToolCommandLineInvoker<T extends BuildToolCommandLine
         } finally {
             mutator.afterScenario(scenarioContext);
         }
+    }
+
+    /**
+     * Returns a {@link Supplier} that returns the result of the given command.
+     */
+    private BuildStepAction<R> measureCommandLineExecution(List<String> commandLine, File workingDir, File buildLog) {
+        return new BuildStepAction<R>() {
+            @Override
+            public boolean isDoesSomething() {
+                return true;
+            }
+
+            @Override
+            public R run(BuildContext buildContext, BuildStep buildStep) {
+                Timer timer = new Timer();
+                if (buildLog == null) {
+                    new CommandExec().inDir(workingDir).run(commandLine);
+                } else {
+                    new CommandExec().inDir(workingDir).runAndCollectOutput(buildLog, commandLine);
+                }
+                Duration executionTime = timer.elapsed();
+                return (R) new BuildInvocationResult(buildContext, executionTime);
+            }
+        };
     }
 }
