+++ src/main/java/org/gradle/profiler/MavenScenarioDefinition.java
@@ -8,7 +8,7 @@ import java.util.function.Supplier;
 public class MavenScenarioDefinition extends BuildToolCommandLineScenarioDefinition {
     public MavenScenarioDefinition(
         String scenarioName,
-        String title,
+        @Nullable String title,
         List<String> targets,
         Supplier<BuildMutator> buildMutator,
         int warmUpCount,
