+++ src/main/java/org/gradle/profiler/ScenarioInvoker.java
@@ -3,13 +3,10 @@ package org.gradle.profiler;
 import org.gradle.profiler.result.BuildInvocationResult;
 import org.gradle.profiler.result.Sample;
 
-import java.io.File;
 import java.io.IOException;
-import java.time.Duration;
 import java.util.Collections;
 import java.util.List;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
 public abstract class ScenarioInvoker<T extends ScenarioDefinition, R extends BuildInvocationResult> {
     /**
@@ -39,29 +36,5 @@ public abstract class ScenarioInvoker<T extends ScenarioDefinition, R extends Bu
         R result = new RunBuildStepAction<R>(action, buildMutator).run(buildContext, BuildStep.BUILD);
         consumer.accept(result);
         return result;
-    }
-
-    /**
-     * Returns a {@link Supplier} that returns the result of the given command.
-     */
-    BuildStepAction<R> measureCommandLineExecution(List<String> commandLine, File workingDir, File buildLog) {
-        return new BuildStepAction<R>() {
-            @Override
-            public boolean isDoesSomething() {
-                return true;
-            }
-
-            @Override
-            public R run(BuildContext buildContext, BuildStep buildStep) {
-                Timer timer = new Timer();
-                if (buildLog == null) {
-                    new CommandExec().inDir(workingDir).run(commandLine);
-                } else {
-                    new CommandExec().inDir(workingDir).runAndCollectOutput(buildLog, commandLine);
-                }
-                Duration executionTime = timer.elapsed();
-                return (R) new BuildInvocationResult(buildContext, executionTime);
-            }
-        };
     }
 }
